requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-shortuuidfield==0.1.3